requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
PyMuPDF
python-docx
nltk
